MovieTicketAPI/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
alembic 
psycopg2-binary 
pydantic[dotenv] 
pinecone
python-dotenv
transformers
torch
passlib
bcrypt